pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-to-editable-ppt"
version = "0.1.0"
description = "Conservative structural diagram to editable PPTX converter"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy>=1.26",
  "Pillow>=11.0",
  "python-pptx>=1.0.2",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10"]

[project.scripts]
image-to-editable-ppt = "image_to_editable_ppt.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
